abstract messages from ConsoleClient to Display class
