+++ app/src/main/java/tttTDD/Java/ConsoleClient.java
@@ -2,13 +2,12 @@ package tttTDD.Java;
 
 import tttTDD.Java.Interfaces.Player;
 
-import java.util.ArrayList;
 import java.util.InputMismatchException;
 import java.util.Scanner;
-import java.util.stream.IntStream;
 
 public class ConsoleClient {
     private PlayerFactory playerFactory = new PlayerFactory();
+    private Display display = new Display();
 
     private Game game;
     private final Scanner input;
@@ -21,14 +20,14 @@ public class ConsoleClient {
     }
 
     public void startGame() {
-        renderBoard();
+        display.renderBoard(game, gameConfig.getBoardSizeConfiguration());
         consoleLoop();
     }
 
     public void consoleLoop() {
         while (!game.isTheGameComplete()) {
             if (!game.getCurrentPlayer().isComputer()) {
-                askPlayerToMakeMove();
+                System.out.println(display.makeMoveMessage());
             }
 
             performPlayerTurn();
@@ -51,17 +50,11 @@ public class ConsoleClient {
 
     private void performPlayerTurn() {
         if (game.performTurn(game.getCurrentPlayerMove())) {
-            printTheMovePerformed(game.getCurrentPlayerPreviousMove());
-            renderBoard();
+            System.out.println(display.showMovePerformedMessage(game, game.getCurrentPlayerPreviousMove()));;
+            System.out.println(display.renderBoard(game, gameConfig.getBoardSizeConfiguration()));;
         } else {
-            askPlayerToPlayAgain();
+            System.out.println(display.tryAgainMessage());
         }
-    }
-
-    private void printTheMovePerformed(int value) {
-        String marker = game.getCurrentPlayer().getMarker().equals(game.getPlayer1Marker()) ? game.getPlayer2Marker() : game.getPlayer1Marker();
-        String message = String.format("Player %s has made a move on space %s", marker, value);
-        System.out.println(message);
     }
 
     private void printGameResult() {
@@ -69,66 +62,8 @@ public class ConsoleClient {
     }
 
     private int askForBoardSize() {
-        System.out.println("Please enter a board size number:");
+        System.out.println(display.enterBoardSizeMessage());
         return boardSizeInput();
-    }
-
-    private void askPlayerToMakeMove() {
-        String message = "Make your move:";
-        System.out.println(message);
-    }
-
-    private void askPlayerToPlayAgain() {
-        String message = "Oops... incorrect move. Please try again";
-        System.out.println(message);
-    }
-
-    private void renderBoard() {
-        int boardSize = gameConfig.getBoardSizeConfiguration();
-        String[] board = game.getBoard();
-        String largestInputOnBoard = String.valueOf(boardSize * boardSize);
-        String separators = drawRowHorizontalLines(boardSize, largestInputOnBoard);
-
-        ArrayList<String> fullBoard = new ArrayList<>();
-        ArrayList<String> formattedRow = new ArrayList<>();
-
-        for(int location = 0; location < board.length; location++) {
-            String cell = formatCellDisplay(board[location], largestInputOnBoard);
-            formattedRow.add(cell);
-
-            int reachedEndOfTheRow = (location + 1) % boardSize;
-
-            if (reachedEndOfTheRow == 0) {
-                String linesBetweenCells = String.join(" |", formattedRow);
-                fullBoard.add(linesBetweenCells + "\n");
-                formattedRow.clear();
-            }
-        }
-
-        String assembleBoard = String.join(separators +"\n",fullBoard);
-
-        System.out.println(assembleBoard);
-    }
-
-    private String drawRowHorizontalLines(int boardSize, String maxInput) {
-        ArrayList<String> separatorCollection = new ArrayList<>();
-
-        String separator = maxInput.replaceAll(".", "-") + "-";
-        IntStream.range(0, boardSize).forEach(i -> separatorCollection.add(separator));
-
-        return String.join("+", separatorCollection);
-    }
-
-    private String formatCellDisplay(String cell, String maxInput) {
-        StringBuilder formattedCell = new StringBuilder();
-
-        if (cell.length() <= maxInput.length()) {
-
-            IntStream.range(0, maxInput.length() - cell.length()).forEach(i -> formattedCell.append(" "));
-            formattedCell.append(cell);
-        }
-
-        return formattedCell.toString();
     }
 
     private int boardSizeInput() {
@@ -143,13 +78,13 @@ public class ConsoleClient {
         }
 
         catch (InputMismatchException e) {
-            System.out.println(wrongTypeMessage());
+            System.out.println(display.wrongTypeMessage());
             input.nextLine();
             return  boardSizeInput();
         }
 
         catch (Exception e) {
-            System.out.println(negativeBoardSize());
+            System.out.println(display.negativeBoardSizeMessage());
             input.nextLine();
             return  boardSizeInput();
         }
@@ -166,20 +101,12 @@ public class ConsoleClient {
         }
 
         catch (Exception e) {
-            System.out.println(wrongTypeMessage());
+            System.out.println(display.wrongTypeMessage());
             input.nextLine();
             return  playerSelectionInput();
         }
 
         return value;
-    }
-
-    private String wrongTypeMessage() {
-        return "Stop it. Enter a number as an input:";
-    }
-
-    private String negativeBoardSize() {
-        return "Can't have the Universe implode on us. Try a larger number:";
     }
 
     private Player askToCreatePlayer(String marker) {
@@ -189,7 +116,8 @@ public class ConsoleClient {
     }
 
     private int askForPlayerType() {
-        System.out.println("What type of player is this?[Enter a number]\n 1. Human\n 2. Computer\n 3. Random Computer\n");
+        System.out.println(display.playerSelectionMessage());
+
         return playerSelectionInput();
     }
 }
